+++ src/app/global-variables.ts
@@ -1,14 +1,17 @@
 
 import { animate, animateChild, group, query, state, style, transition, trigger } from '@angular/animations';
 import { LinkInfo } from "./classes/LinkInfo";
+
 import { InvestmentPageComponent } from './pages/investment-page/investment-page.component';
 import { InvestmentVersusPageComponent } from './pages/investment-versus-page/investment-versus-page.component';
 import { PageProjectsAndExamplesComponent } from './pages/page-projects-and-examples/page-projects-and-examples.component';
+import { KellyCriterionPageComponent } from './pages/kelly-criterion-page/kelly-criterion-page.component';
 
 export const LINK_INFOS: LinkInfo[] = [
   new LinkInfo("Investment", 'investment', InvestmentVersusPageComponent, [
     new LinkInfo("DCA versus VA", 'investment/versus', InvestmentVersusPageComponent), //Dollar Cost Averaging, Value Averaging
     new LinkInfo("Custom DCA versus VA", 'investment/custom', InvestmentPageComponent),
+    new LinkInfo("Kelly Criterion", 'investment/kelly-criterion', KellyCriterionPageComponent),
   ]),
   new LinkInfo("Projects&Examples", 'projects-and-examples', PageProjectsAndExamplesComponent),
 ];
@@ -17,27 +20,23 @@ export const FADE_ANIMATION =
   trigger('routeAnimations', [
     transition('* <=> *', [
       style({ position: 'relative' }),
-      query(':enter, :leave', [
-        style({
-          position: 'absolute',
-          top: 0,
-          left: 0,
-          width: '100%'
-        })
-      ], { optional: true }),
+      group([
+        query(':enter, :leave', [
+          style({
+            'grid-column-start': 1,
+            'grid-row-start': 1,
+          })
+        ], { optional: true }),
+      ]),
       query(':enter', [
         style({ opacity: 0 })
       ], { optional: true }),
-      query(':leave', animateChild(), { optional: true }),
-      group([
-        query(':leave', [
-          animate('500ms ease-out', style({ opacity: 0 }))
-        ], { optional: true }),
-        query(':enter', [
-          animate('500ms ease-out', style({ opacity: 1 }))
-        ], { optional: true })
-      ]),
-      query(':enter', animateChild(), { optional: true }),
+      query(':leave', [
+        animate('300ms ease-out', style({ opacity: 0 }))
+      ], { optional: true }),
+      query(':enter', [
+        animate('500ms ease-out', style({ opacity: 1 }))
+      ], { optional: true }),
     ]),
   ]);